+++ DummyFile.Generator/DummyFile.Generator.Services/Abstract/Item.cs
@@ -2,13 +2,13 @@ namespace DummyFile.Generator.Services.Abstract;
 
 public readonly struct Item
 {
-    public Item(int number, ReadOnlyMemory<byte> text)
+    public Item(long number, ReadOnlyMemory<byte> text)
     {
         Number = number;
         Text = text;
     }
 
-    public int Number { get; }
+    public long Number { get; }
 
     public ReadOnlyMemory<byte> Text { get; }
